test(client-lib): reorganize integration tests infrastructure
By splitting the `fake` module into several files so they can be browsed
more easily.
